pyproject.toml
[project]
name = "matlab-mcp-tools"
version = "0.1.0"
description = "MCP server for MATLAB integration with Cline"
authors = [
    { name = "Yahya" }
]
dependencies = [
    "mcp[cli]>=0.1.0",
    "matlabengine",
    "pydantic>=2.0.0",
]
requires-python = ">=3.8"
readme = "README.md"
license = { file = "LICENSE" }

[project.scripts]
matlab-mcp = "matlab_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/matlab_mcp"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long
]

[tool.ruff.lint.isort]
known-first-party = ["matlab_mcp"]
